refactor(pagination): extract MessageIntervalPaginator base from MessagePaginator
Split the message-interval mechanics (interval store, ingestItem, reflect*, jump navigation,
truncate, focus signals, query/filters/sort/cursor scaffolding) into a new MessageIntervalPaginator
base. MessagePaginator now extends it and adds only the unread/live-view concern
(unreadStateSnapshot, liveViewState, seed/set/clearUnreadSnapshot, isViewingLive/setViewingLive,
jumpToTheFirstUnreadMessage, and the postQueryReconcile/clearStateAndCache overrides).

Pure refactor: MessagePaginator's public surface is unchanged (shared types re-exported), so no
consumer changes. This gives PinnedMessagePaginator an unread-free base to extend by construction.
